requirements.txt
azure-cognitiveservices-nspkg==2.0.0
azure-cognitiveservices-search-imagesearch==1.0.0
azure-cognitiveservices-search-nspkg==2.0.0
azure-common==1.1.16
azure-nspkg==3.0.2
boto3==1.9.23
botocore==1.12.23
certifi==2018.8.24
chardet==3.0.4
Click==7.0
docutils==0.14
Flask==1.0.2
Flask-Cors==3.0.6
gunicorn==19.9.0
idna==2.7
isodate==0.6.0
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.0
mongoengine==0.15.3
msrest==0.6.0
nltk==3.3
oauthlib==2.1.0
pymongo==3.7.2
python-dateutil==2.7.3
requests==2.19.1
requests-oauthlib==1.0.0
s3transfer==0.1.13
six==1.11.0
urllib3==1.23
Werkzeug==0.14.1
